requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.0
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
pycparser==2.21
Pygments==2.17.2
python-dotenv==1.0.0
rich==13.7.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
